requirements.txt
requests
dotenv
pymysql